Albatrosses: add

+++ server/api/task/mutation.ts
@@ -3,8 +3,11 @@ import { queryDB } from "../../entity";
 import { Tasks } from "../../entity/tasks";
 import { getNowString, getNow } from "../../helper";
 import { generateResolver } from "../../helper/log";
+import { generateAuth, verifyAuth } from "../../helper/verify";
+import { MESSAGE_WORD } from "../enum";
+import { roleConfig } from "../config/common";
 
-export const addTask = async (_, { addTaskInput }): Promise<any> => {
+export const addTask = async (_, { addTaskInput }, context): Promise<any> => {
   const {
     name,
     simple,
@@ -19,7 +22,12 @@ export const addTask = async (_, { addTaskInput }): Promise<any> => {
     reviewer
   } = addTaskInput;
 
-  await queryDB(async connection => {
+  return await queryDB(async connection => {
+    const currentUser = await generateAuth(context, connection);
+    if (!verifyAuth(currentUser, "customer")) {
+      return generateResolver(false, MESSAGE_WORD.UNAUTH);
+    }
+
     const task = new Tasks();
     task.name = name;
     task.simple = simple;
@@ -37,29 +45,41 @@ export const addTask = async (_, { addTaskInput }): Promise<any> => {
 
     const tasksRepository = connection.getRepository(Tasks);
 
-    return await tasksRepository.save(task);
+    await tasksRepository.save(task);
+    return generateResolver(true, MESSAGE_WORD.ADD_SUCCESS);
   });
-
-  return generateResolver(true, "添加成功");
 };
 
-export const removeTask = async (_, { removeTaskInput }): Promise<any> => {
+export const removeTask = async (
+  _,
+  { removeTaskInput },
+  context
+): Promise<any> => {
   const { id } = removeTaskInput;
 
   return await queryDB(async connection => {
+    const currentUser = await generateAuth(context, connection);
+    if (!verifyAuth(currentUser, "customer")) {
+      return generateResolver(false, MESSAGE_WORD.UNAUTH);
+    }
+
     const tasksRepository = connection.getRepository(Tasks);
     const task = await tasksRepository.findOne({ id });
 
     if (task) {
       await tasksRepository.remove(task);
-      return generateResolver(true, "删除成功");
+      return generateResolver(true, MESSAGE_WORD.DELETE_SUCCESS);
     } else {
-      return generateResolver(false, "删除失败，任务不存在");
+      return generateResolver(false, MESSAGE_WORD.DELETE_ERROR_TASK_NOT_FOUND);
     }
   });
 };
 
-export const updateTask = async (_, { updateTaskInput }): Promise<any> => {
+export const updateTask = async (
+  _,
+  { updateTaskInput },
+  context
+): Promise<any> => {
   const {
     id,
     name,
@@ -77,6 +97,11 @@ export const updateTask = async (_, { updateTaskInput }): Promise<any> => {
   } = updateTaskInput;
 
   return await queryDB(async connection => {
+    const currentUser = await generateAuth(context, connection);
+    if (!verifyAuth(currentUser, "customer")) {
+      return generateResolver(false, MESSAGE_WORD.UNAUTH);
+    }
+
     const tasksRepository = connection.getRepository(Tasks);
     const task = await tasksRepository.findOne({ id });
 
@@ -95,9 +120,9 @@ export const updateTask = async (_, { updateTaskInput }): Promise<any> => {
       task.updateDate = getNow();
       task.reviewer = reviewer;
       await tasksRepository.save(task);
-      return generateResolver(true, "更新成功");
+      return generateResolver(true, MESSAGE_WORD.UPDATE_SUCCESS);
     } else {
-      return generateResolver(false, "更新失败，任务不存在");
+      return generateResolver(false, MESSAGE_WORD.DELETE_ERROR_TASK_NOT_FOUND);
     }
   });
 };
